requirement.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==1.11.20
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-filter==2.2.0
djangorestframework==3.10.2
idna==2.8
Markdown==3.1.1
mysqlclient==1.4.4
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.3
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
urllib3==1.25.3
